Move adapter loaders & base adapter mixin into separate modules
